refactor: reset the structure of code
